pyproject.toml
[tool.poetry]
name = "thesisv2"
version = "0.1.0"
description = ""
authors = ["wkhoury <wkhoury>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.3"
numpy = "^1.24.2"
faker = "^18.2.0"
faker-music = "^0.4"
configparser = "^5.3.0"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
